+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log (generator-ego)
 
+## 0.34.0
+
+* change `backend*` => `service*`
+
+## 0.33.0
+
+* Redis and frontend features can be selected or deselected now
+
 ## 0.32.0
 
 * re-add `npm install`s in `backend-node-express` generator